unnecessary re-renders on the explore route
