+++ src/components/explore/filter/FilterItemsMenu.jsx
@@ -1,8 +1,24 @@
-import { Blockchain, Category, Status, Price, Dropdown } from "./";
+import { useState } from "react";
+import { Blockchain, Category, Status, Price, Dropdown } from ".";
 import { Dbutton_active_lg } from "./dark/Dbutton_active_lg";
 import { Wbutton_active_lg } from "./white/Wbutton_active_lg";
 
-export const Container = ({ onDropdown, dropdown }) => {
+export const FilterItemsMenu = () => {
+  const [dropdown, setDropdown] = useState({
+    blockchain: false,
+    status: false,
+    price: false,
+    category: false,
+    collections: false,
+  });
+
+  const onDropdown = (type) => {
+    setDropdown({
+      ...dropdown,
+      [type]: !dropdown[type],
+    });
+  };
+
   return (
     <div className="bg-white flex flex-col gap-2 py-3 px-4 flex-grow shrink-0 basis-0 tablet:w-[18.5rem] tablet:basis-auto tablet:py-1 tablet:px-0 tablet:shrink-0 tablet:rounded-2xl tablet:border-light-border tablet:border-[0.7px]">
       <div className="flex flex-col gap-1 py-1 flex-grow shrink-0 basis-0 self-stretch overflow-y-auto tablet:overflow-visible">
