+++ src/components/explore/filter/index.js
@@ -1,7 +1,7 @@
 export * from "./Dropdown";
 export * from "./Filter";
 export * from "./Header";
-export * from "./Container";
+export * from "./FilterItemsMenu";
 export * from "./Blockchain";
 export * from "./Status";
 export * from "./Category";
